fix bugs

+++ src/main/java/com/example/shopee/controller/admin/AdminCategoryController.java
@@ -2,6 +2,7 @@ package com.example.shopee.controller.admin;
 
 import com.example.shopee.entity.CategoryEntity;
 import com.example.shopee.repository.CategoryRepository;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -19,21 +20,36 @@ public class AdminCategoryController {
 
 
     @GetMapping("")
-    public String categoryPage(Model model) {
+    public String categoryPage(Model model, HttpSession session) {
+        String email = (String) session.getAttribute("email");
+
+        if (email == null) {
+            return "redirect:/admin/login";
+        }
         List<CategoryEntity> categories = categoryRepository.findAll();
         model.addAttribute("categories", categories);
         return "admin/category/list";
     }
 
     @GetMapping("/insertCategoryPage")
-    public String insertCategoryPage(Model model) {
+    public String insertCategoryPage(Model model, HttpSession session) {
+        String email = (String) session.getAttribute("email");
+
+        if (email == null) {
+            return "redirect:/admin/login";
+        }
         CategoryEntity categoryEntity = new CategoryEntity();
         model.addAttribute("category", categoryEntity);
         return "admin/category/insert";
     }
 
     @PostMapping("/save")
-    public String save(@ModelAttribute(name = "category") CategoryEntity categoryEntity, Model model) {
+    public String save(@ModelAttribute(name = "category") CategoryEntity categoryEntity, Model model, HttpSession session) {
+        String email = (String) session.getAttribute("email");
+
+        if (email == null) {
+            return "redirect:/admin/login";
+        }
         Optional<CategoryEntity> existingCategory = categoryRepository.findByCategoryName(categoryEntity.getCategoryName());
         if (existingCategory.isPresent()) {
             model.addAttribute("error", "Danh mục này đã tồn tại.");
@@ -45,14 +61,24 @@ public class AdminCategoryController {
     }
 
     @GetMapping("/updateCategory/{id}")
-    public String getFormUpdateCategory(@PathVariable("id") Long id, Model model) {
+    public String getFormUpdateCategory(@PathVariable("id") Long id, Model model, HttpSession session) {
+        String email = (String) session.getAttribute("email");
+
+        if (email == null) {
+            return "redirect:/admin/login";
+        }
         CategoryEntity categoryEntity = categoryRepository.findById(id).get();
         model.addAttribute("category", categoryEntity);
         return "admin/category/update";
     }
 
     @PostMapping("/update")
-    public String update(@ModelAttribute(name = "category") CategoryEntity categoryEntity, Model model) {
+    public String update(@ModelAttribute(name = "category") CategoryEntity categoryEntity, Model model, HttpSession session) {
+        String email = (String) session.getAttribute("email");
+
+        if (email == null) {
+            return "redirect:/admin/login";
+        }
         Optional<CategoryEntity> existingCategory = categoryRepository.findByCategoryName(categoryEntity.getCategoryName());
         if (existingCategory.isPresent() && !existingCategory.get().getId().equals(categoryEntity.getId())) {
             model.addAttribute("error", "Danh mục này đã tồn tại.");
@@ -65,7 +91,12 @@ public class AdminCategoryController {
     }
 
     @GetMapping("/deleteCategory/{id}")
-    public String delete(@PathVariable("id") Long id) {
+    public String delete(@PathVariable("id") Long id, HttpSession session) {
+        String email = (String) session.getAttribute("email");
+
+        if (email == null) {
+            return "redirect:/admin/login";
+        }
         categoryRepository.deleteById(id);
         return "redirect:/admin/category";
     }
